requirements.txt
discord.py==2.3.2
python-dotenv
boto3
botocore
PyNaCl
mutagen 
requests